Implement genre selection callback and update film page layout for dynamic content display
