+++ pages/callbacks.py
@@ -21,30 +21,3 @@ def display_page(pathname):
     return render_page_content(pathname)
 
 
-# Define the callback to update the content based on the button clicked
-@dash.callback(
-    Output("content-area", "children"),
-    [Input("btn-action", "n_clicks"),
-     Input("btn-fantasy", "n_clicks"),
-     Input("btn-drama", "n_clicks"),
-     Input("btn-sci-fi", "n_clicks")]
-)
-
-def update_content(btn_action, btn_fantasy, btn_drama, btn_sci_fi):
-    ctx = dash.callback_context
-
-    if not ctx.triggered:
-        button_id = "btn-action"  # Default to Action content
-    else:
-        button_id = ctx.triggered[0]["prop_id"].split(".")[0]
-
-    if button_id == "btn-action":
-        return html.P("Content for Action genre.")
-    elif button_id == "btn-fantasy":
-        return html.P("Content for Fantasy genre.")
-    elif button_id == "btn-drama":
-        return html.P("Content for Drama genre.")
-    elif button_id == "btn-sci-fi":
-        return html.P("Content for Sci-Fi genre.")
-    else:
-        return html.P("Select a genre to see the content.")
